requirements.txt
django
djangorestframework
gunicorn
whitenoise
dj-database-url
django-cors-headers
psycopg2-binary

PyPDF2
pdfminer.six
spacy

langchain
google-generativeai
